Update: #220 datasource/remote/ai/impl KMP Phase 1 — 서비스 commonMain 이전 + Hilt 모듈 app으로 이관
